refactor(github-issues): extract issue command comment helpers
